#1025 -- display all anchor locations.  #1026 -- display different Anchor.Look values.  Remove unneeded vb.net project and a unneeded cpp file.

+++ GUIs/Editor/FlowChart.cs
@@ -1291,8 +1291,8 @@ namespace SysCAD.Editor
                 //if (state.PortCheck(out requestId, (destinationBox.Tag as Item).Guid, destinationAnchorChosen) == PortStatus.Available)
                 {
                   SysCAD.Protocol.Point anchorPointPos = GetRelativeAnchorPosition(new SysCAD.Protocol.Rectangle(destinationBox.BoundingRect),
-                    (float)((originAnchorChosen.Positions[0] as SysCAD.Protocol.Point).X),
-                    (float)((originAnchorChosen.Positions[0] as SysCAD.Protocol.Point).Y),
+                    (originAnchorChosen.Positions[0] as SysCAD.Protocol.Point).X,
+                    (originAnchorChosen.Positions[0] as SysCAD.Protocol.Point).Y,
                     destinationBox.RotationAngle);
 
                   PointF[] extensionPoints =
